refactor(db): extract slash_command_usage and pinned_command into db/commands.rs
